pyproject.toml
[project]
name = "pydanticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nest-asyncio>=1.6.0",
    "pydantic-ai>=0.1.3",
    "python-dotenv>=1.1.0",
]
